remote_repositories/API service/requirements.txt
fastapi
uvicorn
langchain
pdfminer-six
sentence-transformers
faiss-cpu
tqdm
pandas
